Update sword2.md
A few obvious typos corrected. The English is not always fluent in my ears but I'm not a native speaker.

The service document of sword-v1 points to https://easy.dans.knaw.nl/schemas/docs/sword-v1-packaging.html  I suppose the service document of sword-v2 should point to a similar document, the last time I looked it didn't. The boxed sections of the packaging document are [generated] and I expect them to be identical for both sword interfaces.

Rather than copy-pasting the pdf's from the dans website I would like to see a single location and some automated link checker.

[generated]: https://github.com/DANS-KNAW/easy-schema#maintenance-of-the-packaging-document